% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_reporting_data.R
\name{epidata_source}
\alias{epidata_source}
\title{Create an epidata data source object}
\usage{
epidata_source(signal, geo_types = c("state", "nation"))
}
\arguments{
\item{signal}{Character, epidata signal name (e.g., "confirmed_admissions_covid_ew_prelim")}

\item{geo_types}{Character vector, geographic types to query (e.g., "state", "nation")}
}
\value{
A source object of class "epidata_source"
}
\description{
Create an epidata data source object
}
