% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_reporting_data.R
\name{fetch_reporting_data.default}
\alias{fetch_reporting_data.default}
\title{Default method for unsupported source types}
\usage{
\method{fetch_reporting_data}{default}(source, ...)
}
\description{
Default method for unsupported source types
}
