% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_reporting_data.R
\name{github_source}
\alias{github_source}
\title{Create a GitHub data source object}
\usage{
github_source(
  signal = "covid",
  repo = NULL,
  file_path = NULL,
  github_token = NULL
)
}
\arguments{
\item{signal}{Character, disease signal ("covid", "flu", or "rsv")}

\item{repo}{Character, GitHub repository (e.g., "CDCgov/covid19-forecast-hub"). Defaults based on signal.}

\item{file_path}{Character, path to target data file within repo. Defaults based on signal.}

\item{github_token}{Character, GitHub personal access token for authentication}
}
\value{
A source object of class "github_source"
}
\description{
Create a GitHub data source object
}
