% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcastNHSN-package.R
\docType{package}
\name{nowcastNHSN-package}
\alias{nowcastNHSN}
\alias{nowcastNHSN-package}
\title{nowcastNHSN: NHSN Nowcasting with Reporting Triangles}
\description{
Functions for generating reporting triangles from versioned epidatr data and running nowcasts using baselinenowcast for NHSN hospital admissions.
}
\author{
\strong{Maintainer}: CDC CFA \email{cfa@cdc.gov}

}
\keyword{internal}
