% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_utils.R
\name{saturdays_to_epirange}
\alias{saturdays_to_epirange}
\title{Convert Saturday dates to \code{epidatr::epirange}}
\usage{
saturdays_to_epirange(dates)
}
\arguments{
\item{dates}{A vector of dates. Must all be Saturdays.}
}
\value{
An \code{epidatr::epirange} object covering the epiweeks corresponding
to the input dates.
}
\description{
Converts a sequence of Saturday dates (forecast hub convention) to an
\code{epidatr::epirange} object (Sunday-ending weeks in YYYYWW format).
This allows using the same date inputs for both GitHub fetcher (Saturdays)
and epidatr fetcher (epiweeks).
}
\details{
The function:
\enumerate{
\item Validates that all dates are Saturdays
\item Adds one day to convert to Sunday (epiweeks end on Sunday)
\item Converts to MMWR epiweek format (YYYYWW)
\item Returns an \code{epidatr::epirange} spanning from the earliest to latest week
}

\strong{Why the conversion?}
\itemize{
\item Forecast hub uses Saturday as the week-ending date
\item MMWR/CDC epiweeks end on Sunday
\item epidatr API requires epiweek format (YYYYWW)
}
}
